[Discover] Rearrange the code
